refactor(cli): move tab components to tabs/ subdir, rm -v test
