+++ packages/cli/src/ui/components/tabs/permissions-tab.tsx
