+++ packages/cli/src/ui/components/tabs/prd-items-tab.tsx
